+++ juriscraper/OpinionSite.py
@@ -39,6 +39,7 @@ class OpinionSite(AbstractSite):
             "types",
             "other_dates",
             "attorneys",
+            "headnotes",
         ]
         self._req_attrs = [
             "case_dates",
@@ -136,6 +137,9 @@ class OpinionSite(AbstractSite):
         return None
 
     def _get_attorneys(self):
+        return None
+
+    def _get_headnotes(self):
         return None
 
     def extract_from_text(self, scraped_text: str) -> dict: