+++ CHANGES.md
@@ -25,12 +25,12 @@ Fixes:
 
 ## Current
 
-**2.6.72 - 2025-06-04**
+**2.6.74 - 2025-06-04**
 
 - Add `test_hashes` optional argument to `sample_caller`. Helpful to detect
 timestamped opinions and check if `Site.cleanup_content` is working #1392
 - fix tenn scraper parsing error #1413
-
+- fix package release process #1426
 
 **2.6.71 - 2025-05-30**